--- .idea/modules/app/Ziririt.app.iml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module version="4">
-  <component name="ModuleClassLoaderOverlays">
-    <paths>
-      <option value="/var/folders/vk/c_sbyk9d5w54qx4226rc295h0000gn/T/overlay5000993649970280513" />
-      <option value="/var/folders/vk/c_sbyk9d5w54qx4226rc295h0000gn/T/overlay7685327711258521962" />
-      <option value="/var/folders/vk/c_sbyk9d5w54qx4226rc295h0000gn/T/overlay941520690604120565" />
-      <option value="/var/folders/vk/c_sbyk9d5w54qx4226rc295h0000gn/T/overlay12713189597188641071" />
-    </paths>
-  </component>
-</module>
